Move log shipping code to the lib module
