+++ lib/cronitor.go
@@ -2,8 +2,9 @@ package lib
 
 import (
 	"bytes"
+	"compress/gzip"
 	"encoding/json"
-	"errors"
+	"github.com/pkg/errors"
 	"fmt"
 	"github.com/getsentry/raven-go"
 	"github.com/spf13/viper"
@@ -225,4 +226,95 @@ func (api CronitorApi) sendHttpGet(url string) ([]byte, error) {
 	}
 
 	return contents, nil
+}
+
+func gzipLogData(logData string) *bytes.Buffer {
+	var b bytes.Buffer
+	if len(logData) < 1 {
+		return &b
+	}
+
+	gz := gzip.NewWriter(&b)
+	if _, err := gz.Write([]byte(logData)); err != nil {
+		//log("error writing gzip")
+		return nil
+	}
+	if err := gz.Close(); err != nil {
+		//log("error closing gzip")
+		return nil
+	}
+	return &b
+}
+
+func getPresignedUrl(apiKey string, postBody []byte) ([]byte, error) {
+	url := "https://cronitor.io/api/logs/presign"
+
+	client := &http.Client{Timeout: 120 * time.Second}
+	request, err := http.NewRequest("POST", url, strings.NewReader(string(postBody)))
+	if err != nil {
+		return nil, errors.Wrap(err, "could not create request for URL presign")
+	}
+	request.SetBasicAuth(apiKey, "")
+	request.Header.Add("Content-Type", "application/json")
+	response, err := client.Do(request)
+	if err != nil {
+		return nil, errors.Wrap(err, "error requesting presigned url")
+	}
+	if response.StatusCode != 200 && response.StatusCode != 201 {
+		return nil, fmt.Errorf("error response code %d returned", response.StatusCode)
+	}
+
+	contents, err := ioutil.ReadAll(response.Body)
+	if err != nil {
+		return nil, err
+	}
+	defer response.Body.Close()
+	response.Body = ioutil.NopCloser(bytes.NewBuffer(contents))
+	return contents, nil
+}
+
+func SendLogData(apiKey string, monitorKey string, seriesID string, outputLogs string) ([]byte, error) {
+	gzippedLogs := gzipLogData(outputLogs)
+	jsonBytes, err := json.Marshal(map[string]string{
+		"job_key": monitorKey,
+		"series":  seriesID,
+	})
+	if err != nil {
+		return nil, errors.Wrap(err, "couldn't encode job and series IDs to JSON")
+	}
+	var responseJson struct {
+		Url string `json:"url"`
+	}
+	response, err := getPresignedUrl(apiKey, jsonBytes)
+	if err != nil {
+		return nil, errors.Wrap(err, "error generating presign url for log uploading")
+	}
+	if err := json.Unmarshal(response, &responseJson); err != nil {
+		return nil, err
+	}
+	s3LogPutUrl := responseJson.Url
+	if len(s3LogPutUrl) == 0 {
+		return nil, errors.New("no presigned S3 url returned. Something is wrong")
+	}
+	req, err := http.NewRequest("PUT", s3LogPutUrl, gzippedLogs)
+	if err != nil {
+		return nil, err
+	}
+	client := &http.Client{
+		Timeout: 120 * time.Second,
+	}
+	response2, err := client.Do(req)
+	if err != nil || response == nil {
+		return nil, errors.Wrap(err, fmt.Sprintf("error putting logs: %v", response2))
+	}
+	if response2.StatusCode < 200 || response2.StatusCode >= 300 {
+		return nil, fmt.Errorf("error response code %d returned", response2.StatusCode)
+	}
+	body, err := ioutil.ReadAll(response2.Body)
+	if err != nil {
+		return nil, err
+	}
+	defer response2.Body.Close()
+	//log(fmt.Sprintf("logs shipped for series %s", seriesID))
+	return body, nil
 }
